feat: add global var to enable static checking
It does nothing yet since the typechecker is not ready.

But the idea is that the developer can toggle this var on REPL
when he wants to enable/disable static checking

Also the ideia is to call a function to typecheck a specific function,
namespace or whole project

Maybe this function can be added to a CI/CD pipeline, so the static
checking is moved away from developing time, but still assured when
opening a pull request.